LayoutTests/http/tests/download/anchor-download-redirect-same-origin-top-level-expected.txt

Download started.
Download was redirected to "http://127.0.0.1:8000/download/resources/sample-render.html".
Downloading URL with suggested filename "sample-render.html"
Download completed.
Download
